+++ addons/storysource/src/register.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import addons from '@storybook/addons';
+import { addons } from '@storybook/addons';
 
 import { StoryPanel } from './StoryPanel';
 import { ADDON_ID, PANEL_ID } from '.';