+++ lib/client-api/src/client_api.ts
@@ -80,12 +80,28 @@ export const addArgTypesEnhancer = (enhancer: ArgTypesEnhancer) => {
   singleton.addArgTypesEnhancer(enhancer);
 };
 
+export const getGlobalRender = () => {
+  if (!singleton)
+    throw new Error(`Singleton client API not yet initialized, cannot call getGlobalRender`);
+
+  return singleton.globalRender;
+};
+
+export const setGlobalRender = (render: StoryFn) => {
+  if (!singleton)
+    throw new Error(`Singleton client API not yet initialized, cannot call setGobalRender`);
+  singleton.globalRender = render;
+};
+
+const invalidStoryTypes = new Set(['string', 'number', 'boolean', 'symbol']);
 export default class ClientApi {
   private _storyStore: StoryStore;
 
   private _addons: ClientApiAddons<unknown>;
 
   private _decorateStory: DecorateStoryFunction;
+
+  private _globalRender: StoryFn<any>;
 
   // React Native Fast refresh doesn't allow multiple dispose calls
   private _noStoryModuleAddMethodHotDispose: boolean;
@@ -151,6 +167,14 @@ export default class ClientApi {
   addArgTypesEnhancer = (enhancer: ArgTypesEnhancer) => {
     this._storyStore.addArgTypesEnhancer(enhancer);
   };
+
+  get globalRender(): StoryFn {
+    return this._globalRender;
+  }
+
+  set globalRender(render: StoryFn) {
+    this._globalRender = render;
+  }
 
   // what are the occasions that "m" is a boolean vs an obj
   storiesOf = <StoryFnReturnType = unknown>(
@@ -218,9 +242,9 @@ export default class ClientApi {
         throw new Error(`Invalid or missing storyName provided for a "${kind}" story.`);
       }
 
-      if (typeof storyFn !== 'function') {
+      if (!storyFn || Array.isArray(storyFn) || invalidStoryTypes.has(typeof storyFn)) {
         throw new Error(
-          `Cannot load story "${storyName}" in "${kind}" due to invalid format. Storybook expected a function but received ${typeof storyFn} instead.`
+          `Cannot load story "${storyName}" in "${kind}" due to invalid format. Storybook expected a function/object but received ${typeof storyFn} instead.`
         );
       }