+++ lib/client-api/src/story_store.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-underscore-dangle */
 import createChannel from '@storybook/channel-postmessage';
 import { toId } from '@storybook/csf';
-import addons, { mockChannel } from '@storybook/addons';
+import { addons, mockChannel } from '@storybook/addons';
 import Events from '@storybook/core-events';
 
 import StoryStore from './story_store';
@@ -46,6 +46,7 @@ const addStoryToStore = (store, kind, name, storyFn, parameters = {}) =>
       id: toId(kind, name),
     },
     {
+      // FIXME: need applyHooks, but this breaks the current tests
       applyDecorators: defaultDecorateStory,
     }
   );
@@ -145,8 +146,8 @@ describe('preview.story_store', () => {
       const story = jest.fn();
       addStoryToStore(store, 'a', '1', story);
 
-      const { getDecorated } = store.getRawStory('a', '1');
-      getDecorated()();
+      const context = store.getRawStory('a', '1');
+      context.getDecorated()(context);
 
       expect(globalDecorator).toHaveBeenCalled();
       expect(kindDecorator).toHaveBeenCalled();