+++ lib/core-common/src/types.ts
@@ -1,5 +1,5 @@
 import type ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-import type { PluginOptions } from '@storybook/react-docgen-typescript-plugin';
+import type { PluginOptions } from 'react-docgen-typescript-plugin';
 import { Configuration, Stats } from 'webpack';
 import { TransformOptions } from '@babel/core';
 import { Router } from 'express';
@@ -155,6 +155,7 @@ export interface BuilderOptions {
   cache: FileSystemCache;
   configDir: string;
   docsMode: boolean;
+  previewCsfV3?: boolean;
   versionCheck?: VersionCheck;
   releaseNotesData?: ReleaseNotesData;
   disableWebpackDefaults?: boolean;
@@ -235,12 +236,21 @@ export interface StorybookConfig {
   >;
   core?: CoreConfig;
   logLevel?: string;
-  /**
-   * Allows to disable deprecated implicit PostCSS loader.
-   */
   features?: {
+    /**
+     * Allows to disable deprecated implicit PostCSS loader.
+     */
     postcss?: boolean;
+
+    /**
+     * Build stories.json automatically on start/build
+     */
     buildStoriesJson?: boolean;
+
+    /**
+     * Activate preview of CSF v3.0
+     */
+    previewCsfV3?: boolean;
   };
   /**
    * Tells Storybook where to find stories.