+++ lib/csf-tools/src/CsfFile.ts
@@ -9,6 +9,7 @@ import { toId, isExportStory, storyNameFromExport } from '@storybook/csf';
 const logger = console;
 interface Meta {
   title?: string;
+  component?: string;
   includeStories?: string[] | RegExp;
   excludeStories?: string[] | RegExp;
 }
@@ -40,23 +41,8 @@ const parseTitle = (value: any) => {
   return undefined;
 };
 
-const parseMeta = (declaration: t.ObjectExpression): Meta => {
-  const meta: Meta = {};
-  declaration.properties.forEach((p: Node) => {
-    if (t.isObjectProperty(p) && t.isIdentifier(p.key)) {
-      if (p.key.name === 'title') {
-        meta.title = parseTitle(p.value);
-      } else if (['includeStories', 'excludeStories'].includes(p.key.name)) {
-        // @ts-ignore
-        meta[p.key.name] = parseIncludeExclude(p.value);
-      }
-    }
-  });
-  return meta;
-};
-
 const findVarInitialization = (identifier: string, program: t.Program) => {
-  let init: t.Node = null;
+  let init: t.Expression = null;
   let declarations: t.VariableDeclarator[] = null;
   program.body.find((node: t.Node) => {
     if (t.isVariableDeclaration(node)) {
@@ -83,7 +69,7 @@ const findVarInitialization = (identifier: string, program: t.Program) => {
   return init;
 };
 
-const isArgsStory = ({ init }: t.VariableDeclarator, parent: t.Node) => {
+const isArgsStory = ({ init }: t.VariableDeclarator, parent: t.Node, csf: CsfFile) => {
   let storyFn: t.Node = init;
   // export const Foo = Bar.bind({})
   if (t.isCallExpression(init)) {
@@ -102,6 +88,8 @@ const isArgsStory = ({ init }: t.VariableDeclarator, parent: t.Node) => {
       const boundIdentifier = callee.object.name;
       const template = findVarInitialization(boundIdentifier, parent);
       if (template) {
+        // eslint-disable-next-line no-param-reassign
+        csf._templates[boundIdentifier] = template;
         storyFn = template;
       }
     }
@@ -112,14 +100,45 @@ const isArgsStory = ({ init }: t.VariableDeclarator, parent: t.Node) => {
   return false;
 };
 export class CsfFile {
-  _ast: Node;
+  _ast: t.File;
 
   _meta?: Meta;
 
   _stories: Record<string, Story> = {};
 
-  constructor(ast: Node) {
+  _metaAnnotations: Record<string, Node> = {};
+
+  _storyExports: Record<string, t.VariableDeclarator> = {};
+
+  _storyAnnotations: Record<string, Record<string, Node>> = {};
+
+  _templates: Record<string, t.Expression> = {};
+
+  constructor(ast: t.File) {
     this._ast = ast;
+  }
+
+  _parseMeta(declaration: t.ObjectExpression) {
+    const meta: Meta = {};
+    declaration.properties.forEach((p: t.ObjectProperty) => {
+      if (t.isIdentifier(p.key)) {
+        this._metaAnnotations[p.key.name] = p.value;
+
+        if (p.key.name === 'title') {
+          meta.title = parseTitle(p.value);
+        } else if (['includeStories', 'excludeStories'].includes(p.key.name)) {
+          // @ts-ignore
+          meta[p.key.name] = parseIncludeExclude(p.value);
+        } else if (p.key.name === 'component') {
+          if (t.isIdentifier(p.value)) {
+            meta.component = p.value.name;
+          } else if (t.isStringLiteral(p.value)) {
+            meta.component = p.value.value;
+          }
+        }
+      }
+    });
+    this._meta = meta;
   }
 
   parse() {
@@ -128,20 +147,25 @@ export class CsfFile {
     traverse(this._ast, {
       ExportDefaultDeclaration: {
         enter({ node, parent }) {
+          let metaNode: t.ObjectExpression;
           if (t.isObjectExpression(node.declaration)) {
             // export default { ... };
-            self._meta = parseMeta(node.declaration);
+            metaNode = node.declaration;
           } else if (
             // export default { ... } as Meta<...>
             t.isTSAsExpression(node.declaration) &&
             t.isObjectExpression(node.declaration.expression)
           ) {
-            self._meta = parseMeta(node.declaration.expression);
+            metaNode = node.declaration.expression;
           } else if (t.isIdentifier(node.declaration) && t.isProgram(parent)) {
             const init = findVarInitialization(node.declaration.name, parent);
             if (t.isObjectExpression(init)) {
-              self._meta = parseMeta(init);
+              metaNode = init;
             }
+          }
+
+          if (!self._meta && metaNode) {
+            self._parseMeta(metaNode);
           }
         },
       },
@@ -154,51 +178,91 @@ export class CsfFile {
                 const { name } = decl.id;
                 const parameters = {
                   // __id: toId(self._meta.title, name),
-                  // FiXME: Template.bind({});
-                  __isArgsStory: isArgsStory(decl, parent),
+                  // FIXME: Template.bind({});
+                  __isArgsStory: isArgsStory(decl, parent, self),
                 };
                 self._stories[name] = {
                   id: 'FIXME',
                   name,
                   parameters,
                 };
+                self._storyExports[name] = decl;
+                if (self._storyAnnotations[name]) {
+                  logger.warn(`Unexpected annotations for "${name}" before story declaration`);
+                } else {
+                  self._storyAnnotations[name] = {};
+                }
               }
             });
           }
         },
       },
       ExpressionStatement: {
-        enter({ node }) {
+        enter({ node, parent }) {
           const { expression } = node;
           // B.storyName = 'some string';
           if (
+            t.isProgram(parent) &&
             t.isAssignmentExpression(expression) &&
             t.isMemberExpression(expression.left) &&
             t.isIdentifier(expression.left.object) &&
-            t.isIdentifier(expression.left.property, { name: 'storyName' }) &&
-            t.isStringLiteral(expression.right)
+            t.isIdentifier(expression.left.property)
           ) {
             const exportName = expression.left.object.name;
-            const storyName = expression.right.value;
-            const story = self._stories[exportName];
-            if (!story) return;
-            story.name = storyName;
+            const annotationKey = expression.left.property.name;
+            const annotationValue = expression.right;
+
+            // v1-style annotation
+            // A.story = { parameters: ..., decorators: ... }
+
+            if (self._storyAnnotations[exportName]) {
+              if (annotationKey === 'story' && t.isObjectExpression(annotationValue)) {
+                annotationValue.properties.forEach((prop: t.ObjectProperty) => {
+                  if (t.isIdentifier(prop.key)) {
+                    self._storyAnnotations[exportName][prop.key.name] = prop.value;
+                  }
+                });
+              } else {
+                self._storyAnnotations[exportName][annotationKey] = annotationValue;
+              }
+            } else {
+              logger.debug(`skipping "${exportName}.${annotationKey}"`);
+            }
+
+            if (annotationKey === 'storyName' && t.isStringLiteral(annotationValue)) {
+              const storyName = annotationValue.value;
+              const story = self._stories[exportName];
+              if (!story) return;
+              story.name = storyName;
+            }
           }
         },
       },
     });
 
     // default export can come at any point in the file, so we do this post processing last
-    self._stories =
-      self._meta && self._meta.title
-        ? Object.entries(self._stories).reduce((acc, [key, story]) => {
-            if (isExportStory(key, self._meta)) {
-              const id = toId(self._meta.title, storyNameFromExport(key));
-              acc[key] = { ...story, id, parameters: { ...story.parameters, __id: id } };
-            }
-            return acc;
-          }, {} as Record<string, Story>)
-        : {}; // no meta = no stories
+    if (self._meta?.title || self._meta?.component) {
+      self._stories = Object.entries(self._stories).reduce((acc, [key, story]) => {
+        if (isExportStory(key, self._meta)) {
+          const id = toId(self._meta.title, storyNameFromExport(key));
+          acc[key] = { ...story, id, parameters: { ...story.parameters, __id: id } };
+        }
+        return acc;
+      }, {} as Record<string, Story>);
+
+      Object.keys(self._storyExports).forEach((key) => {
+        if (!isExportStory(key, self._meta)) {
+          delete self._storyExports[key];
+          delete self._storyAnnotations[key];
+        }
+      });
+    } else {
+      // no meta = no stories
+      self._stories = {};
+      self._storyExports = {};
+      self._storyAnnotations = {};
+    }
+
     return self;
   }
 
@@ -211,7 +275,7 @@ export class CsfFile {
   }
 }
 
-export const loadCsf = async (code: string) => {
+export const loadCsf = (code: string) => {
   const ast = parse(code, {
     sourceType: 'module',
     // FIXME: we should get this from the project config somehow?
@@ -225,7 +289,7 @@ export const loadCsf = async (code: string) => {
   return new CsfFile(ast);
 };
 
-export const formatCsf = async (csf: CsfFile) => {
+export const formatCsf = (csf: CsfFile) => {
   const { code } = generate(csf._ast, {});
   return code;
 };